start modal to modify  booking
